+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 # 0.1.4
 * Define `Member` datatype for chat users
 * No implicit `Prelude` anymore (NoImplicitPrelude)
+* Move HTTP request function into `Internal` module